+++ src/tests/test_console.py
@@ -1,18 +1,24 @@
+import asyncio
 import pytest
 from pathlib import Path
-import time
 import sys
 
+from pyodide_build.testing import run_in_pyodide
 from conftest import selenium_common
 
 sys.path.append(str(Path(__file__).resolve().parents[2] / "src" / "py"))
 
-from pyodide import console, CodeRunner  # noqa: E402
-from pyodide.console import _CodeRunnerCompile, _CodeRunnerCommandCompiler  # noqa: E402
+from pyodide import CodeRunner  # noqa: E402
+from _pyodide.console import (
+    Console,
+    _Compile,
+    _CommandCompiler,
+)  # noqa: E402
+from _pyodide import console
 
 
 def test_command_compiler():
-    c = _CodeRunnerCompile()
+    c = _Compile()
     with pytest.raises(SyntaxError, match="unexpected EOF while parsing"):
         c("def test():\n   1", "<input>", "single")
     assert isinstance(c("def test():\n   1\n", "<input>", "single"), CodeRunner)
@@ -23,7 +29,7 @@ def test_command_compiler():
     )
     assert isinstance(c("1<>2", "<input>", "single"), CodeRunner)
 
-    c = _CodeRunnerCommandCompiler()
+    c = _CommandCompiler()
     assert c("def test():\n   1", "<input>", "single") is None
     assert isinstance(c("def test():\n   1\n", "<input>", "single"), CodeRunner)
     with pytest.raises(SyntaxError, match="invalid syntax"):
@@ -34,7 +40,7 @@ def test_command_compiler():
     assert isinstance(c("1<>2", "<input>", "single"), CodeRunner)
 
 
-def test_stream_redirection():
+def test_write_stream():
     my_buffer = ""
 
     def callback(string):
@@ -49,114 +55,7 @@ def test_stream_redirection():
     assert my_buffer == "foo\nbar\n"
 
 
-@pytest.fixture
-def safe_sys_redirections():
-    redirected = sys.stdout, sys.stderr, sys.displayhook
-    try:
-        yield
-    finally:
-        sys.stdout, sys.stderr, sys.displayhook = redirected
-
-
-def test_interactive_console_streams(safe_sys_redirections):
-    my_stdout = ""
-    my_stderr = ""
-    orig_sys_stdout_name = sys.stdout.name
-    orig_sys_stderr_name = sys.stderr.name
-
-    def stdout_callback(string):
-        nonlocal my_stdout
-        my_stdout += string
-
-    def stderr_callback(string):
-        nonlocal my_stderr
-        my_stderr += string
-
-    ##########################
-    # Persistent redirection #
-    ##########################
-    shell = console._InteractiveConsole(
-        stdout_callback=stdout_callback,
-        stderr_callback=stderr_callback,
-        persistent_stream_redirection=True,
-    )
-
-    # std names
-    assert sys.stdout.name == orig_sys_stdout_name
-    assert sys.stderr.name == orig_sys_stderr_name
-
-    # std redirections
-    print("foo")
-    assert my_stdout == "foo\n"
-    print("bar", file=sys.stderr)
-    assert my_stderr == "bar\n"
-
-    shell.push("print('foobar')")
-    assert my_stdout == "foo\nfoobar\n"
-
-    shell.push("print('foobar')")
-    assert my_stdout == "foo\nfoobar\nfoobar\n"
-
-    shell.push("1+1")
-    assert my_stdout == "foo\nfoobar\nfoobar\n2\n"
-    assert shell.run_complete.result() == 2
-
-    my_stderr = ""
-    shell.push("raise Exception('hi')")
-    assert (
-        my_stderr
-        == 'Traceback (most recent call last):\n  File "<console>", line 1, in <module>\nException: hi\n'
-    )
-    assert shell.run_complete.exception() is not None
-    my_stderr = ""
-    shell.push("1+1")
-    assert my_stderr == ""
-    assert shell.run_complete.result() == 2
-
-    del shell
-    import gc
-
-    gc.collect()
-
-    my_stdout = ""
-    my_stderr = ""
-
-    print("bar")
-    assert my_stdout == ""
-
-    print("foo", file=sys.stdout)
-    assert my_stderr == ""
-
-    ##############################
-    # Non persistent redirection #
-    ##############################
-    shell = console._InteractiveConsole(
-        stdout_callback=stdout_callback,
-        stderr_callback=stderr_callback,
-        persistent_stream_redirection=False,
-    )
-
-    print("foo")
-    assert my_stdout == ""
-
-    shell.push("print('foobar')")
-    assert my_stdout == "foobar\n"
-
-    print("bar")
-    assert my_stdout == "foobar\n"
-
-    shell.push("print('foobar')")
-    assert my_stdout == "foobar\nfoobar\n"
-
-    shell.push("import sys")
-    shell.push("print('foobar', file=sys.stderr)")
-    assert my_stderr == "foobar\n"
-
-    shell.push("1+1")
-    assert my_stdout == "foobar\nfoobar\n2\n"
-
-
-def test_repr(safe_sys_redirections):
+def test_repr():
     sep = "..."
     for string in ("x" * 10 ** 5, "x" * (10 ** 5 + 1)):
         for limit in (9, 10, 100, 101):
@@ -165,89 +64,9 @@ def test_repr(safe_sys_redirections):
             ) == 2 * (limit // 2) + len(sep)
 
 
-@pytest.fixture
-def safe_selenium_sys_redirections(selenium):
-    # Import console early since it makes three global hiwire allocations, and we don't want to anger
-    # the memory leak checker
-    selenium.run_js("pyodide._module.runPythonSimple(`from pyodide import console`)")
-
-    selenium.run_js(
-        "pyodide._module.runPythonSimple(`import sys; _redirected = sys.stdout, sys.stderr, sys.displayhook`)"
-    )
-    try:
-        yield
-    finally:
-        selenium.run_js(
-            "pyodide._module.runPythonSimple(`sys.stdout, sys.stderr, sys.displayhook = _redirected`)"
-        )
-
-
-def test_interactive_console(selenium, safe_selenium_sys_redirections):
-    selenium.run(
-        """
-        import sys
-        from pyodide.console import _InteractiveConsole
-
-        result = None
-
-        def display(value):
-            global result
-            result = value
-
-        shell = _InteractiveConsole()
-        shell.display = display
-        """
-    )
-
-    selenium.run("shell.push('x = 5')")
-    selenium.run("shell.push('x')")
-    selenium.run_js("await pyodide.runPythonAsync('await shell.run_complete');")
-    assert selenium.run("result") == 5
-
-    selenium.run("shell.push('x ** 2')")
-    selenium.run_js("await pyodide.runPythonAsync('await shell.run_complete');")
-
-    assert selenium.run("result") == 25
-
-    selenium.run("shell.push('def f(x):')")
-    selenium.run("shell.push('    return x*x + 1')")
-    selenium.run("shell.push('')")
-    selenium.run("shell.push('str([f(x) for x in range(5)])')")
-    selenium.run_js("await pyodide.runPythonAsync('await shell.run_complete');")
-    assert selenium.run("result") == str([1, 2, 5, 10, 17])
-
-    selenium.run("shell.push('def factorial(n):')")
-    selenium.run("shell.push('    if n < 2:')")
-    selenium.run("shell.push('        return 1')")
-    selenium.run("shell.push('    else:')")
-    selenium.run("shell.push('        return n * factorial(n - 1)')")
-    selenium.run("shell.push('')")
-    selenium.run("shell.push('factorial(10)')")
-    selenium.run_js("await pyodide.runPythonAsync('await shell.run_complete');")
-    assert selenium.run("result") == 3628800
-
-    # with package load
-    selenium.run("shell.push('import pytz')")
-    selenium.run("shell.push('pytz.utc.zone')")
-    selenium.run_js("await pyodide.runPythonAsync('await shell.run_complete');")
-    assert selenium.run("result") == "UTC"
-
-
-def test_completion(selenium, safe_selenium_sys_redirections):
-    selenium.run(
-        """
-        from pyodide import console
-
-        shell = console._InteractiveConsole()
-        """
-    )
-
-    assert selenium.run(
-        """
-        [completions, start] = shell.complete('a')
-        [tuple(completions), start]
-        """
-    ) == [
+def test_completion():
+    shell = Console({"a_variable": 7})
+    shell.complete("a") == (
         [
             "and ",
             "as ",
@@ -258,48 +77,219 @@ def test_completion(selenium, safe_selenium_sys_redirections):
             "all(",
             "any(",
             "ascii(",
+            "a_variable",
         ],
         0,
-    ]
+    )
 
-    assert selenium.run(
-        """
-        [completions, start] = shell.complete('a = 0 ; print.__g')
-        [tuple(completions), start]
-        """
-    ) == [
+    assert shell.complete("a = 0 ; print.__g") == (
         [
             "print.__ge__(",
             "print.__getattribute__(",
             "print.__gt__(",
         ],
         8,
-    ]
-
-
-def test_interactive_console_top_level_await(selenium, safe_selenium_sys_redirections):
-    selenium.run(
-        """
-        import sys
-        from pyodide.console import _InteractiveConsole
-
-        result = None
-
-        def display(value):
-            global result
-            result = value
-
-        shell = _InteractiveConsole()
-        shell.display = display
-        """
     )
-    selenium.run("shell.push('from js import fetch')")
-    time.sleep(0.2)
-    selenium.run("""shell.push("await (await fetch('packages.json')).json()")""")
-    time.sleep(0.2)
-    res = selenium.run("result")
-    assert isinstance(res, dict)
-    assert res["dependencies"]["micropip"] == ["pyparsing", "packaging", "distutils"]
+
+
+def test_interactive_console():
+    shell = Console()
+
+    def assert_incomplete(input):
+        res = shell.push(input)
+        assert res.syntax_check == "incomplete"
+
+    async def get_result(input):
+        res = shell.push(input)
+        assert res.syntax_check == "complete"
+        return await res
+
+    async def test():
+        assert await get_result("x = 5") == None
+        assert await get_result("x") == 5
+        assert await get_result("x ** 2") == 25
+
+        assert_incomplete("def f(x):")
+        assert_incomplete("    return x*x + 1")
+        assert await get_result("") == None
+        assert await get_result("[f(x) for x in range(5)]") == [1, 2, 5, 10, 17]
+
+        assert_incomplete("def factorial(n):")
+        assert_incomplete("    if n < 2:")
+        assert_incomplete("        return 1")
+        assert_incomplete("    else:")
+        assert_incomplete("        return n * factorial(n - 1)")
+        assert await get_result("") == None
+        assert await get_result("factorial(10)") == 3628800
+
+        assert await get_result("import pytz") == None
+        assert await get_result("pytz.utc.zone") == "UTC"
+
+        fut = shell.push("1+")
+        assert fut.syntax_check == "syntax-error"
+        assert fut.exception() is not None
+        assert (
+            fut.formatted_error
+            == '  File "<console>", line 1\n    1+\n      ^\nSyntaxError: invalid syntax\n'
+        )
+
+        fut = shell.push("raise Exception('hi')")
+        try:
+            await fut
+        except:
+            assert (
+                fut.formatted_error
+                == 'Traceback (most recent call last):\n  File "<console>", line 1, in <module>\nException: hi\n'
+            )
+
+    asyncio.get_event_loop().run_until_complete(test())
+
+
+def test_top_level_await():
+    from asyncio import Queue, sleep
+
+    q = Queue()
+    shell = Console(locals())
+    fut = shell.push("await q.get()")
+
+    async def test():
+        await sleep(0.3)
+        assert not fut.done()
+        await q.put(5)
+        assert await fut == 5
+
+    asyncio.get_event_loop().run_until_complete(test())
+
+
+@pytest.fixture
+def safe_sys_redirections():
+    redirected = sys.stdout, sys.stderr, sys.displayhook
+    try:
+        yield
+    finally:
+        sys.stdout, sys.stderr, sys.displayhook = redirected
+
+
+def test_persistent_redirection(safe_sys_redirections):
+    my_stdout = ""
+    my_stderr = ""
+    orig_stdout = sys.stdout
+    orig_stderr = sys.stderr
+
+    def stdout_callback(string):
+        nonlocal my_stdout
+        my_stdout += string
+
+    def stderr_callback(string):
+        nonlocal my_stderr
+        my_stderr += string
+
+    shell = Console(
+        stdout_callback=stdout_callback,
+        stderr_callback=stderr_callback,
+        persistent_stream_redirection=True,
+    )
+
+    # std names
+    assert sys.stdout.name == orig_stdout.name
+    assert sys.stderr.name == orig_stderr.name
+
+    # std redirections
+    print("foo")
+    assert my_stdout == "foo\n"
+    print("bar", file=sys.stderr)
+    assert my_stderr == "bar\n"
+    my_stderr = ""
+
+    async def get_result(input):
+        res = shell.push(input)
+        assert res.syntax_check == "complete"
+        return await res
+
+    async def test():
+        assert await get_result("print('foobar')") == None
+        assert my_stdout == "foo\nfoobar\n"
+
+        assert await get_result("print('foobar')") == None
+        assert my_stdout == "foo\nfoobar\nfoobar\n"
+
+        assert await get_result("1+1") == 2
+        assert my_stdout == "foo\nfoobar\nfoobar\n"
+
+    asyncio.get_event_loop().run_until_complete(test())
+
+    my_stderr = ""
+
+    shell.persistent_restore_streams()
+    my_stdout = ""
+    my_stderr = ""
+    print(sys.stdout, file=orig_stdout)
+    print("bar")
+    assert my_stdout == ""
+
+    print("foo", file=sys.stdout)
+    assert my_stderr == ""
+
+
+def test_nonpersistent_redirection(safe_sys_redirections):
+    my_stdout = ""
+    my_stderr = ""
+
+    def stdout_callback(string):
+        nonlocal my_stdout
+        my_stdout += string
+
+    def stderr_callback(string):
+        nonlocal my_stderr
+        my_stderr += string
+
+    async def get_result(input):
+        res = shell.push(input)
+        assert res.syntax_check == "complete"
+        return await res
+
+    shell = Console(
+        stdout_callback=stdout_callback,
+        stderr_callback=stderr_callback,
+        persistent_stream_redirection=False,
+    )
+
+    print("foo")
+    assert my_stdout == ""
+
+    async def test():
+        assert await get_result("print('foobar')") == None
+        assert my_stdout == "foobar\n"
+
+        print("bar")
+        assert my_stdout == "foobar\n"
+
+        assert await get_result("print('foobar')") == None
+        assert my_stdout == "foobar\nfoobar\n"
+
+        assert await get_result("import sys") == None
+        assert await get_result("print('foobar', file=sys.stderr)") == None
+        assert my_stderr == "foobar\n"
+
+        assert await get_result("1+1") == 2
+
+    asyncio.get_event_loop().run_until_complete(test())
+
+
+@pytest.mark.skip_refcount_check
+@run_in_pyodide
+async def test_console_imports():
+    from pyodide.console import PyodideConsole
+
+    shell = PyodideConsole()
+
+    async def get_result(input):
+        res = shell.push(input)
+        assert res.syntax_check == "complete"
+        return await res
+
+    assert await get_result("import pytz") == None
+    assert await get_result("pytz.utc.zone") == "UTC"
 
 
 @pytest.fixture(params=["firefox", "chrome"], scope="function")
@@ -349,14 +339,14 @@ SyntaxError: invalid syntax]`
         await term.ready;
         result.push([term.get_output(),
 `>>> raise Exception('hi')
-[[;;;terminal-error]Traceback (most recent call last):]
-[[;;;terminal-error]  File "<console>", line 1, in <module>]
-[[;;;terminal-error]Exception: hi]`
+[[;;;terminal-error]Traceback (most recent call last):
+  File "<console>", line 1, in <module>
+Exception: hi]`
         ]);
 
         term.clear();
         term.exec("from _pyodide_core import trigger_fatal_error; trigger_fatal_error()");
-        await term.ready;
+        await sleep(100);
         result.push([term.get_output(),
 `>>> from _pyodide_core import trigger_fatal_error; trigger_fatal_error()
 [[;;;terminal-error]Pyodide has suffered a fatal error. Please report this to the Pyodide maintainers.]
@@ -365,7 +355,6 @@ SyntaxError: invalid syntax]`
 [[;;;terminal-error]Look in the browser console for more details.]`
         ]);
 
-        await sleep(30);
         assert(() => term.paused());
         return result;
         """