+++ src/tests/test_pyodide.py
@@ -258,8 +258,8 @@ def test_run_python_async_toplevel_await(selenium):
         await pyodide.runPythonAsync(`
             from js import fetch
             resp = await fetch("packages.json")
-            json = await resp.json()
-            assert hasattr(json, "dependencies")
+            json = (await resp.json()).to_py()["packages"]
+            assert "micropip" in json
         `);
         """
     )
@@ -514,8 +514,10 @@ def test_fatal_error(selenium_standalone):
         x = re.sub("/lib/python.*/", "", x)
         x = re.sub("/lib/python.*/", "", x)
         x = re.sub("warning: no [bB]lob.*\n", "", x)
-        x = re.sub("Error: intentionally triggered fatal error!", "{}", x)
+        x = re.sub("Error: intentionally triggered fatal error!\n", "", x)
         x = re.sub(" +at .*\n", "", x)
+        x = re.sub(".*@https?://[0-9.:]*/.*\n", "", x)
+        x = x.replace("\n\n", "\n")
         return x
 
     assert (
@@ -526,7 +528,6 @@ def test_fatal_error(selenium_standalone):
                 Python initialization complete
                 Pyodide has suffered a fatal error. Please report this to the Pyodide maintainers.
                 The cause of the fatal error was:
-                {}
                 Stack (most recent call first):
                   File "<exec>", line 8 in h
                   File "<exec>", line 6 in g